refactor(fields): give the dynamic-field machinery its own module
forms.py had crossed a thousand lines and was already a grab-bag — SVG
validation, panel settings, discounts, spreadsheet regexes. The cluster this
branch kept growing is one concept with one caller-facing purpose, and the
three modules that import it import exactly it: dynamic_fields.py.

Two holes closed on the way. A field slugged "triggers_custom" beside one
slugged "triggers" had its answer merged into the wrong field, so both sides
now leave a real key alone. And `super().clean() or {}` was a fallback that
could only ever fire by returning `{}` for Django to assign over cleaned_data,
wiping it; the pair is read from cleaned_data directly.

The panel keeps the required marker on a select the write-in can answer —
tessera_field takes the requirement rather than reading it off a field whose
own flag we deliberately cleared.

The importer no longer shreds an option that contains a comma even when other
answers follow it: a run of parts that could still grow into a configured
option is held back until it matches one or cannot. `split_answers` says what
that is, next to `split_custom`, which is still about write-ins.
