+++ src/ludamus/templates/panel/parts/proposal-session-fields.html
@@ -7,7 +7,7 @@
             <div class="card-body">
                 <h3 class="text-lg font-semibold text-foreground mb-2">{% translate "Session Fields" %}</h3>
                 {% for desc in field_descriptors %}
-                    {% tessera_field desc.bound_field %}
+                    {% tessera_field desc.bound_field required=desc.is_required %}
                     {% if desc.custom_bound_field %}
                         {% tessera_field desc.custom_bound_field %}
                     {% endif %}
